Zero vector (z) is a vector in which each element is zero.
A zero vector or null vector is a vector that has a zero magnitude and no direction. Before getting into the detail about the zero vector, let us first recall the meaning of vectors. A vector is an object or a geometric entity that has a magnitude (length) and direction. Now, a vector that has zero length and an undefined direction is called a zero vector or null vector. It has all components equal to 0.